Successful connection on the database. it was possible to save and consult data, thanks to small corrections in the code and the use of the data.dao layer

+++ zbank-api/src/main/java/com/zbank/dto/PerfilDTO.java
@@ -11,7 +11,7 @@ public final class PerfilDTO {
     private String nombre;
     private String apellido;
     private TipoDocumentoDTO tipoDocumento;
-    private long numeroDocumento;
+    private int numeroDocumento;
     private DivisaDTO divisa;
     private String nombreUsuario;
     private String clave;
@@ -30,7 +30,7 @@ public final class PerfilDTO {
     }
 
     public PerfilDTO(final UUID id, final String nombre, final String apellido,
-                     final TipoDocumentoDTO tipoDocumento, final long numeroDocumento,
+                     final TipoDocumentoDTO tipoDocumento, final int numeroDocumento,
                      final DivisaDTO divisa, final String nombreUsuario,
                      final String clave, final String correo) {
         setId(id);
@@ -64,7 +64,7 @@ public final class PerfilDTO {
         return tipoDocumento;
     }
 
-    public final long getNumeroDocumento() {
+    public final int getNumeroDocumento() {
         return numeroDocumento;
     }
 
@@ -105,7 +105,7 @@ public final class PerfilDTO {
         return this;
     }
 
-    public final PerfilDTO setNumeroDocumento(final long numeroDocumento) {
+    public final PerfilDTO setNumeroDocumento(final int numeroDocumento) {
         this.numeroDocumento = numeroDocumento;
         return this;
     }
